2019 07 rewritten to use coroutine IntComputer

+++ src/main/kotlin/no/rodland/advent_2019/Util.kt
@@ -22,13 +22,13 @@ fun String.readFirstLineConvertToInts(): List<Int> {
     return this.readFile()[0].map { it.toString().toInt() }
 }
 
-fun permute(range: IntRange): List<List<Int>> {
+fun permute(range: IntRange): List<List<Long>> {
     return range.flatMap { p0 ->
         range.flatMap { p1 ->
             range.flatMap { p2 ->
                 range.flatMap { p3 ->
                     range.map { p4 ->
-                        listOf(p0, p1, p2, p3, p4)
+                        listOf(p0.toLong(), p1.toLong(), p2.toLong(), p3.toLong(), p4.toLong())
                     }
                 }
             }
